Prevent double-loading of Cucumber rake task
Because Bootsy is an Engine, it's cucumber.rake was being loaded twice.
Once for the Engine, then again for the application itself.

This causes `rake cucumber` or just `rake` to cause the cucumber suite
to run twice.

Work around this by invoking a NameError in the context of a running
Rails application.